Update the foreign predictor interface, make generic composer.
The foreign predictors are now generically accepting a dataframe
and list of target/conditions columns from that df to train on,
before the FPs were relying on the underlying dataset heavily.

The sat_metamodel has been changed to composer. The FP composer
is a generic composer that no longer relies on the number and type
of foreign predictors.